+++ src/containers/Form/CapacityRules/__tests__/validation.test.js
@@ -5,22 +5,25 @@ describe('formValidation', () => {
   it('returns proper object when required fields are provided', () => {
     const values = new Map({
       name: 'mockName',
-      type: 'mockType',
     });
-    expect(formValidation(values)).toMatchSnapshot();
+    expect(
+      formValidation(values, { intl: { formatMessage: ({ defaultMessage }) => defaultMessage } })
+    ).toMatchSnapshot();
   });
   it('returns proper object (errors) when required fields are not provided', () => {
     const values = new Map({
       name: '',
-      type: '',
     });
-    expect(formValidation(values)).toMatchSnapshot();
+    expect(
+      formValidation(values, { intl: { formatMessage: ({ defaultMessage }) => defaultMessage } })
+    ).toMatchSnapshot();
   });
   it('returns proper object (errors) when wrong values are provided', () => {
     const values = new Map({
       name: undefined,
-      type: null,
     });
-    expect(formValidation(values)).toMatchSnapshot();
+    expect(
+      formValidation(values, { intl: { formatMessage: ({ defaultMessage }) => defaultMessage } })
+    ).toMatchSnapshot();
   });
 });