+++ src/containers/Form/CapacityRules/layout.js
@@ -10,29 +10,63 @@
 
 import React from 'react';
 import PropTypes from 'prop-types';
-import { DetailHeader, InputField } from 'cx-ui-components';
+import { DetailHeader, InputField, SelectField } from 'cx-ui-components';
 
-export default function CapacityRulesForm({ handleSubmit, isSaving, inherited, userHasUpdatePermission, key }) {
+export default function CapacityRulesForm({
+  handleSubmit,
+  isSaving,
+  inherited,
+  userHasUpdatePermission,
+  key,
+  initialValues,
+  versions,
+  intl: { formatMessage },
+}) {
   return (
     <form onSubmit={handleSubmit} key={key}>
-      <DetailHeader text="Details" />
+      <DetailHeader
+        text={formatMessage({
+          id: 'userProfile.details',
+          defaultMessage: 'Details',
+        })}
+      />
       <InputField
         name="name"
-        label="Name *"
+        label={`${formatMessage({
+          id: 'identityProviders.details.name',
+          defaultMessage: 'Name',
+        })} *`}
         componentType="input"
         inputType="text"
-        placeholder="Name"
+        placeholder={formatMessage({
+          id: 'identityProviders.details.name',
+          defaultMessage: 'Name',
+        })}
         data-automation="CapacityRulesFormFieldName"
         disabled={isSaving || inherited || !userHasUpdatePermission}
       />
-      <InputField
-        name="description"
-        label="Description"
-        componentType="textarea"
-        inputType="text"
-        data-automation="CapacityRulesFormFieldDescription"
-        disabled={isSaving || inherited || !userHasUpdatePermission}
-      />
+      {initialValues.get('id') && (
+        <SelectField
+          name="activeVersion"
+          label={`${formatMessage({
+            id: 'capacityRules.details.version',
+            defaultMessage: 'Version',
+          })}`}
+          disabled={isSaving || inherited || !userHasUpdatePermission || !versions.length}
+          data-automation="CapacityRulesFormFieldActiveVersion"
+          options={
+            versions &&
+            versions.map(({ numericOrderVersion, name, version }) => ({
+              label: `${numericOrderVersion} - ${name}`,
+              value: version,
+            }))
+          }
+          placeholder={formatMessage({
+            id: 'capacityRules.details.version.placeholder',
+            defaultMessage: 'Select a version',
+          })}
+        />
+      )}
     </form>
   );
 }
@@ -43,4 +77,7 @@ CapacityRulesForm.propTypes = {
   isSaving: PropTypes.bool,
   inherited: PropTypes.bool,
   userHasUpdatePermission: PropTypes.bool,
+  versions: PropTypes.array,
+  initialValues: PropTypes.object.isRequired,
+  intl: PropTypes.object.isRequired,
 };