+++ src/redux/epics.js
@@ -29,6 +29,7 @@ import * as contactLayoutsEpics from './modules/entities/contactLayouts/epics';
 import * as contactAttributesEpics from './modules/entities/contactAttributes/epics';
 import * as forecastDashboardsEpics from './modules/reporting/forecastDashboards/epics';
 import * as userProfileEpics from './modules/entities/userProfile/epics';
+import * as capacityRulesEpics from './modules/entities/capacityRules/epics';
 
 export const rootEpic = combineEpics(
   ...Object.values(brandingEpics),
@@ -60,5 +61,6 @@ export const rootEpic = combineEpics(
   ...Object.values(contactLayoutsEpics),
   ...Object.values(contactAttributesEpics),
   ...Object.values(forecastDashboardsEpics),
-  ...Object.values(userProfileEpics)
+  ...Object.values(userProfileEpics),
+  ...Object.values(capacityRulesEpics)
 );