+++ CHANGELOG.md
@@ -9,7 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Beta User Acceptance Testing]
 - IdentityProviders
-
+## [0.96.0] - 2021-01-20
+### Added
+- CXV1-18711 - Config2 - Capacity Rules | Side panel details table
+- CXV1-18712 - Config2 - Capacity Rules | mainEntityForm
 ## [0.95.5] - 2021-01-21
 ### Fixed 
 - CXV1-23700 - Early access features page fails to load if branding style is null